graphicsview:  - [fixed] arrows are pointing into the wrong direction
and other fixes

+++ src/AutomateRoot.cpp
@@ -49,7 +49,7 @@ unsigned int AutomateRoot::getObjectType() {
 }
 
 void AutomateRoot::removeChild( unsigned int index ) {
-  if ( childItems.size() < index ) {
+  if ( (unsigned int)childItems.size() < index ) {
     qDebug() << "Fatal error, childItems.size() < index!";
     qDebug() << "having " << childItems.size() << " childs";
     exit( 0 );
